Update version to 0.0.2. Update dependencies.

+++ rust/src/api/simple.rs
@@ -40,7 +40,7 @@ pub enum RoonEvent {
 
 pub struct BrowseItems {
     pub list: BrowseList,
-    pub offset: usize,
+    pub offset: u32,
     pub items: Vec<BrowseItem>,
 }
